refactor: Split financial-accounting service into logical modules
Break down the 1765-line financial_accounting_service.go monolith into
focused endpoint files, each under 600 lines:

- financial_accounting_service.go (278 lines): Service struct, interfaces,
  constructor, options
- grpc_posting_endpoints.go (556 lines): CaptureLedgerPosting,
  RetrieveLedgerPosting, UpdateLedgerPosting, validation helpers
- grpc_booking_endpoints.go (492 lines): InitiateFinancialBookingLog,
  RetrieveFinancialBookingLog, UpdateFinancialBookingLog, state transitions
- grpc_control_endpoints.go (289 lines): ControlFinancialBookingLog,
  extractUserFromContext
- grpc_ledger_endpoints.go (221 lines): ListFinancialBookingLogs,
  ListLedgerPostings

All code is moved without modification. No behavioral changes.
